style(Certificate): conform more with style rfc
