--- src/Certificate/constants.js
@@ -1,19 +1,0 @@
-// answer is the index of the correct option
-export const certificate_form = [
-  {
-    question: 'As a registered user, I can:',
-    options: ['Browse public Project', 'upload file'],
-    id: 'abc',
-    answer: 0,
-    hint: 'some information about this question',
-  },
-  {
-    question: 'How can I share data',
-    id: 'def',
-    options: ['I can not share data', 'I can only share data with BPA memebers'],
-    answer: 1,
-    hint: 'some information about this question',
-  },
-];
-
-export const title = 'Please complete following questions';
